Refator: 룰렛 캔버스 컴포넌트 화
